refactor: 2S.7-pre: extract ConfigPostProcessor from FileConfigStore.load()
Pure refactor — moves route ordering, deployment-id uniqueness check,
entity-name back-fill, and ToolSet name validation into a new
ConfigPostProcessor class. FileConfigStore.load() shrinks to: parse
config, run ConfigPostProcessor.process(), call apiKeyStore (still
inline; 2S.6-pre relocates), publish via onReloadCallbacks. Behaviour
unchanged; existing FileConfigStoreTest passes.

The slice register row's two-pass split + cross-entity slash-name
rejection are deferred to 2S.9 per its broader cross-entity
validation scope (amended in §5).

Design anchors: 07 Phase 2 prereqs (line 113); 02 §4.1, §9
Tests: server/src/test/java/com/epam/aidial/core/server/config/FileConfigStoreTest.java (existing)
