+++ docs/sandbox/dial-unified-config/IMPLEMENTATION.md
@@ -354,14 +354,14 @@ These map 1:1 to the named prerequisite PRs in `07-migration-and-rollout.md` §P
 | **2S.4-pre** | ~~`ResourceService.put(descriptor, body, skipLock=true)` package-visible overload.~~ **Dropped 2026-05-03** — `ResourceService.putResource(descriptor, body, etag, author, boolean lock)` already exists publicly (`storage/.../ResourceService.java:552`); `lock=false` provides the documented skipLock semantics. Already used externally (`PublicationService.java:337`). | — | 07 Phase 2 prereqs; 04 §2.5 | ❌ dropped | — |
 | **2S.5-pre** | `FileConfigStore` constructor accepts `List<Consumer<Config>> initialOnReloadCallbacks`; registered before `vertx.setPeriodic` to close registration race. | — | 07 Phase 2 prereqs; 02 §4 | ✅ | `fad28222` |
 | **2S.6-pre** | Make `FileConfigStore.load() → apiKeyStore.addProjectKeys` call conditional on `apiKeyStore != null`. Move authoritative call into `ConfigPostProcessor` invoked by `MergedConfigStore`. Wire `MergedConfigStore` to construct `FileConfigStore` with `apiKeyStore = null`. | 2S.5-pre | 07 Phase 2 prereqs; 02 §4 | 📋 | — |
-| **2S.7-pre** | Extract `ConfigPostProcessor` from `FileConfigStore.load()`. Two-pass: structural (always fatal-to-entity) → semantic (skip\|abort per setting). Slash-keyed-name rejection (warn + drop). | — | 07 Phase 2 prereqs; 02 §4.1, §9 | 📋 | — |
+| **2S.7-pre** | Extract `ConfigPostProcessor` from `FileConfigStore.load()` (pure refactor; structural pass only — two-pass split + slash-keyed-name rejection deferred to **2S.9** per its broader cross-entity validation scope). | — | 07 Phase 2 prereqs; 02 §4.1, §9 | ✅ | — |
 
 **Track A — Server core**
 
 | ID | Slice | Depends on | Design anchors | Status | Commit |
 |---|---|---|---|---|---|
 | **2S.8** | `MergedConfigStore` — union of `FileConfigStore` + `ResourceService`. `requestRebuild()` non-blocking entry point. `volatile boolean initialized` guard for pre-init no-op. | 2S.5-pre, 2S.6-pre, 2S.7-pre | 02 §4 | 📋 | — |
-| **2S.9** | Invalid-entity sibling store. Listing/get response shape with `status` + `validationWarnings`. `config.reload.onInvalidEntity: skip\|abort` setting (default `abort` — opt-in `skip` enables the sibling store / status surface). Prometheus `dial_config_skipped_entities`, `dial_config_skip_events_total`. | 2S.8 | 02 §4.1, §4.3; 03 §4 | 📋 | — |
+| **2S.9** | Invalid-entity sibling store. Listing/get response shape with `status` + `validationWarnings`. `config.reload.onInvalidEntity: skip\|abort` setting (default `abort` — opt-in `skip` enables the sibling store / status surface). Prometheus `dial_config_skipped_entities`, `dial_config_skip_events_total`. **Absorbs from 2S.7-pre (2026-05-03):** `ConfigPostProcessor` two-pass split (structural always-fatal → semantic skip\|abort) and slash-keyed-name rejection (warn + drop) across models / applications / interceptors / roles / routes / toolsets — these features are the natural fit for 2S.9's cross-entity validation scope. | 2S.8 | 02 §4.1, §4.3; 03 §4 | 📋 | — |
 | **2S.10** | `SecretFieldProcessor` + `@EncryptedField` annotation in `:config`. Dual `ObjectMapper` (blob I/O vs API response). Mask `***` on Public-view; preserve-on-omit (and `***` sentinel) on `PUT`. Reuses `CredentialEncryptionService` primitives. | — | 04 §2.4–2.6 | 📋 | — |
 | **2S.11** | `MODEL` `ResourceTypes` entry. `POST /v1/models/public/{name}` (409 on conflict). `PUT /v1/models/public/{name}` (404 on missing, optional `If-Match` → 412). `DELETE`. Strict POST/PUT split. Bucket-aware authz. ETag in response header. | 2S.1-pre, 2S.2-pre, 2S.8, 2S.9, 2S.10 | 03 §1, §3; 07 Phase 2 | 📋 | — |
 | **2S.12** | `POST /v1/admin/validate` — model-scoped (Phase 4 extends to other types and bulk). | 2S.11 | 03 §6 | 📋 | — |
